+++ cdk/lib/lambdas/S3FilesLambda/S3FilesPutLambda.ts
@@ -1,13 +1,12 @@
 import { Effect, PolicyStatement } from 'aws-cdk-lib/aws-iam';
 import {Function, Runtime, Code } from 'aws-cdk-lib/aws-lambda';
-import {join} from 'path';
 import {CdkStack} from '../../cdk-stack';
 
 export const S3FilesPutLambda = (parent: CdkStack, s3BucketName: string) => {
     const putLambda = new Function(parent, `S3FilesPutLambda`, {
         runtime: Runtime.NODEJS_12_X,
-        handler: 'index.handler',
-        code: Code.fromAsset(join(__dirname, 'lambda-put-handler')),
+        handler: 's3Index.putS3Handler',
+        code: Code.fromAsset('src'),
         environment: {
             s3Bucket: s3BucketName
         }
