+++ cdk/lib/lambdas/healthLambda/healthLambda.ts
@@ -5,7 +5,7 @@ import {CdkStack} from '../../cdk-stack';
 export const HealthLambda = (parent: CdkStack) : Function => {
     return new Function(parent, 'HealthFunction', {
         runtime: Runtime.NODEJS_12_X,
-        handler: 'index.healthHandler',
+        handler: 'restIndex.healthHandler',
         code: Code.fromAsset('src'),
     });
 }
