--- cdk/lib/lambdas/healthLambda/lambda-handler/index.js
@@ -1,8 +1,0 @@
-exports.handler = async (event) => {
-    // TODO implement
-    const response = {
-        statusCode: 200,
-        body: JSON.stringify('You have reached the test endpoint. Wallahi Neyah Eh!'),
-    };
-    return response;
-};
