+++ cdk/package.json
@@ -9,7 +9,7 @@
     "watch": "tsc -w",
     "test": "jest",
     "cdk": "cdk",
-    "deploy": "cdk deploy",
+    "deploy": "cdk deploy && npm run clean",
     "clean": "tsc --build --clean"
   },
   "devDependencies": {
